Moved is_embed to utils
